Split resolver function value tests
